requirements.txt
asgiref==3.2.10
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1
django-multiselectfield==0.1.12
gunicorn==19.6.0
python-http-client==2.2.1
sendgrid==3.6.3
whitenoise==3.2.3
instamojo-wrapper==1.1
dj-database-url==0.4.2